feat: put upload logic in its own component + reusue it in documentLibraryViewPage
