+++ frontend/src/warfare/pages/TheaterOfOperations.tsx
@@ -31,7 +31,6 @@ export const Geomap = () => {
       <TopBar
         title="Theater of Operations"
         description="Navigate the current theater of operations from data captured in the Mediterranean Sea near the maritime neighbourhood TL."
-        leftLg={4}
       />
 
       {isLoading && <div>Loading map data...</div>}
